requirements.txt
python-dotenv>=0.19.0
groq
AppOpener
pywhatkit
bs4
pillow
rich
requests
keyboard
cohere
googlesearch-python
selenium
mtranslate
pygame
edge-tts
PyQt5
webdriver-manager
textblob>=0.17.1
numpy>=1.21.0
SpeechRecognition>=3.9.0
nltk>=3.6.0